Add reports endpoints to v1 routes.py
